Move 'common' to 'includes'
