+++ drivers/SDM230/driver.js
@@ -1,5 +1,5 @@
 'use strict';
 
-const driver = require('../../common/v1/driver.js');
+const driver = require('../../includes/v1/driver.js');
 
 module.exports = driver;
