+++ drivers/energy_v2/driver.js
@@ -1,6 +1,6 @@
 'use strict';
 
 const Homey = require('homey');
-const driver = require('../../common/v2/Driver');
+const driver = require('../../includes/v2/Driver');
 
 module.exports = driver;
