+++ drivers/energylink/driver.js
@@ -5,7 +5,7 @@ const Homey = require('homey');
 // const { ManagerDrivers } = require('homey');
 // const driver = ManagerDrivers.getDriver('homewizard');
 
-const homewizard = require('../../includes/homewizard.js');
+const homewizard = require('../../includes/legacy/homewizard.js');
 
 let homewizard_devices;
 const devices = {};
