+++ drivers/heatlink/driver.js
@@ -5,7 +5,7 @@ const Homey = require('homey');
 // const { ManagerDrivers } = require('homey');
 // const driver = ManagerDrivers.getDriver('homewizard');
 const devices = {};
-const homewizard = require('../../includes/homewizard.js');
+const homewizard = require('../../includes/legacy/homewizard.js');
 
 let homewizard_devices;
 
@@ -107,7 +107,7 @@ class HomeWizardHeatlink extends Homey.Driver {
 module.exports = HomeWizardHeatlink;
 
 // var devices = {};
-// var homewizard = require('./../../includes/homewizard.js');
+// var homewizard = require('./../../includes/legacy/homewizard.js');
 // var refreshIntervalId = 0;
 //
 // // SETTINGS
