+++ drivers/windmeter/driver.js
@@ -3,7 +3,7 @@
 const Homey = require('homey');
 
 const devices = {};
-const homewizard = require('../../includes/homewizard.js');
+const homewizard = require('../../includes/legacy/homewizard.js');
 
 let homewizard_devices;
