+++ includes/legacy/homewizard.js
